<WIP> 相机朝向四元数乘法bug

+++ OpenGL/src/Camera/Camera3DController.h
@@ -20,7 +20,7 @@ public:
 	float rotateSpeed = 0.1f;
 
 private:
-	float m_cursorPosX;
-	float m_cursorPosY;
+	double m_cursorPosX;
+	double m_cursorPosY;
 
 };
